Add eleMap function to Matrix.prototype
eleMap applies a transformation to each element in a matrix, given its
initial value, its row position, and its columnal position. This
transformation is given to eleMap in the form of a callback function
with parameters v, row, and col.

Typical use cases would be to perform different transformations to
different rows and/or columns. For example, multiply the first row by 2,
the second by 3, and the rest by four. Here is a callback that would
achieve this:

var callback = function(v, row, col) {
    switch (row) {
        case 0: return v * 2;
        case 1: return v * 3;
        default: return v * 4;
    }
}

We would then pass the callback into eleMap as you would into map.

matrix2 = matrix.eleMap(callback);