normalize require and simplify index
Currently, requiring the library returns a function, which must be called to access the Vector and Matrix classes, like:

const { Vector, Matrix } = require('linear-algebra')();

The documentation in quotes above the linearAlgebra function says that the library is supposed to return an Object. I've converted the linearAlgebra export to an immediately invoked function in this commit so that it does return an Object on requiring, and so that requiring can be done as in most JS libraries, like:

const { Vector, Matrix } = require('linear-algebra'); 

Also, absorbed precision() and normal() function calls into the IIFE, which gets rid of some vars.